+++ src/main/java/com/bankinc/api/services/productType/ProductTypeServiceimpl.java
@@ -1,9 +1,11 @@
 package com.bankinc.api.services.productType;
 
+import com.bankinc.api.models.dto.ProductTypeDto;
 import com.bankinc.api.models.entity.TblProductType;
+import com.bankinc.api.models.mappers.ProductTypeMapper;
 import com.bankinc.api.repository.ProductTypeRepository;
-import com.bankinc.api.services.productType.ProductTypeService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,29 +14,36 @@ import java.util.Optional;
 public class ProductTypeServiceimpl implements ProductTypeService {
     @Autowired
     private ProductTypeRepository productTypeRepository;
+    @Autowired
+    private ProductTypeMapper productTypeMapper;
 
     @Override
-    public List<TblProductType> finAll() {
-        return productTypeRepository.findAll();
+    public List<ProductTypeDto> findAll() {
+        return productTypeMapper.toDtos(productTypeRepository.findAll());
     }
 
     @Override
-    public TblProductType save(TblProductType tblProductType) {
-        return productTypeRepository.save(tblProductType);
+    public ProductTypeDto save(TblProductType tblProductType) {
+        return productTypeMapper.toDto(productTypeRepository.save(tblProductType));
     }
 
     @Override
-    public Optional<TblProductType> findById(int id) {
-        return Optional.empty();
+    public Optional<ProductTypeDto> findById(int id) {
+        return productTypeRepository.findById(id)
+                .map(tblProductType -> productTypeMapper.toDto(tblProductType));
     }
 
     @Override
     public String deleteById(int id) {
-        try {
-            productTypeRepository.deleteById(id);
-            return "cliente eliminado";
-        }catch (Exception e){
-            return "fallo en la eliminacion";
+        if(productTypeRepository.existsById(id)){
+            try {
+                productTypeRepository.deleteById(id);
+                return "Cliente eliminado con éxito";
+            } catch (DataAccessException e) {
+                return "Fallo en la eliminación: " + e.getMessage();
+            }
+        } else {
+            return "Cliente no encontrado";
         }
     }
 }
